refactor(console): 입출력 메소드들을 Reader와 Printer 클래스에서 정의
기존 Terminal 클래스에 정의했던 입출력 메소드들을 분리해서 정의
